security-testing-demo/requirements.txt
# Security Testing Dependencies

# ZAP API Client
python-owasp-zap-v2.4>=0.0.21

# Testing Framework
pytest>=7.4.0
pytest-html>=4.1.0
pytest-cov>=4.1.0

# HTTP Requests
requests>=2.31.0

# Report Generation
jinja2>=3.1.0

# Environment Variables
python-dotenv>=1.0.0

# Logging
colorlog>=6.7.0

# Data Validation
jsonschema>=4.20.0
